+++ src/hooks/processManager/state.ts
@@ -5,5 +5,6 @@ const taskManager:RecoilState<TaskType[]> = atom({
   key: 'taskManager',
   default: [] as any
 })
-
+// This atom will be used to manage the state of the task manager
+// It will hold an array of TaskType objects, which represent the tasks in the task manager                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
 export {taskManager}
