+++ src/services/windowManager.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState, Suspense, lazy } from 'react';
 import styled from "styled-components";
 import Discover from "../applications/discover.tsx";
-import Observer from "../applications/utility/Observer.tsx";
+import Observer from "../applications/utility/observer/index.tsx";
 import {useProcessManager} from "../hooks/processManager/processManager.tsx";
 import {TaskType} from "../modules/typeModule.tsx";
 import LogIn from '@/applications/utility/login';
